// components/CardTemplate/card.type.tsx
import { HexColorString } from "@/lib/generic.type";

export type CardConstantsType = {
  height: number;
  width: number;
  bottomPanelPadding: number;
  outerBorderWidth: number;
  innerCornerRadius: number;
  cornerRadius: number;
  hRuleThickness: number;
  outerBorderColour: HexColorString;
  boostCircleRadius: number;
};